backend/requirements.txt
fastapi
uvicorn
groq
torch
torchvision
pydantic
python-multipart
reportlab
numpy
opencv-python-headless
pillow
python-dotenv
langgraph
langchain-groq
ultralytics
opencv-python
